+++ lib/ooxml_parser/common_parser/common_data/alternate_content/drawing/graphic/shape/docx_shape/ooxml_text_box.rb
@@ -16,7 +16,7 @@ module OoxmlParser
         when 'p'
           text_body.elements << DocxParagraph.new.parse(text_body_node_child, index, parent: text_body)
         when 'tbl'
-          text_body.elements << Table.parse(text_body_node_child, index, parent: text_body)
+          text_body.elements << Table.new(parent: text_body).parse(text_body_node_child, index)
         when 'bodyPr'
           text_body.properties = OOXMLShapeBodyProperties.new(parent: text_body).parse(text_body_node_child)
         end